refactor: 컨트롤러,서비스,테스트 클래스 정리
