+++ src/main/resources/schema.sql
@@ -7,12 +7,11 @@ CREATE TABLE STATION
     PRIMARY KEY (id)
 );
 
+DROP TABLE IF EXISTS LINE;
 
-DROP TABLE IF EXISTS LINE2;
-
-CREATE TABLE LINE2
+CREATE TABLE LINE
 (
-    id   BIGINT NO NULL AUTO_INCREMENT,
+    id   BIGINT      NOT NULL AUTO_INCREMENT,
     name VARCHAR(20) NOT NULL,
     PRIMARY KEY (id)
 );
@@ -21,23 +20,24 @@ DROP TABLE IF EXISTS EDGE;
 
 CREATE TABLE EDGE
 (
-    id            BIGINT NOT NULL AUTO_INCREMENT,
-    line_id       BIGINT NOT NULL,
-    station_id    BIGINT NOT NULL,
-    station_order BIGINT NOT NULL,
+    id               BIGINT NOT NULL AUTO_INCREMENT,
+    line_id          BIGINT NOT NULL,
+    station_id       BIGINT NOT NULL,
+    station_order    BIGINT NOT NULL,
+    distance_to_next BIGINT NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (station_id) REFERENCES STATION (id),
-    FOREIGN KEY (line_id) REFERENCES LINE2 (id)
-)
-
-DROP TABLE IF EXISTS LINE;
-
-CREATE TABLE LINE
-(
-    id            BIGINT      NOT NULL AUTO_INCREMENT,
-    name          VARCHAR(20) NOT NULL,
-    station_id    BIGINT      NOT NULL,
-    station_order BIGINT      NOT NULL,
-    FOREIGN KEY (station_id) REFERENCES STATION (id),
-    PRIMARY KEY (id)
+    FOREIGN KEY (line_id) REFERENCES LINE (id)
 );
+
+-- DROP TABLE IF EXISTS LINE;
+--
+-- CREATE TABLE LINE
+-- (
+--     id            BIGINT      NOT NULL AUTO_INCREMENT,
+--     name          VARCHAR(20) NOT NULL,
+--     station_id    BIGINT      NOT NULL,
+--     station_order BIGINT      NOT NULL,
+--     FOREIGN KEY (station_id) REFERENCES STATION (id),
+--     PRIMARY KEY (id)
+-- );
