+++ src/test/java/subway/controller/LineControllerTest.java
@@ -1,0 +1,47 @@
+//package subway.ui;
+//
+//import com.fasterxml.jackson.databind.ObjectMapper;
+//import org.junit.jupiter.api.DisplayName;
+//import org.junit.jupiter.api.Test;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+//import org.springframework.boot.test.mock.mockito.MockBean;
+//import org.springframework.http.MediaType;
+//import org.springframework.test.web.servlet.MockMvc;
+//import subway.application.LineService;
+//import subway.dto.LineCreateRequest;
+//import subway.dto.LineResponse3;
+//
+//import static org.mockito.ArgumentMatchers.any;
+//import static org.mockito.Mockito.when;
+//import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+//import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+//
+//@WebMvcTest(LineController.class)
+//class LineControllerTest {
+//
+//    @Autowired
+//    private MockMvc mockMvc;
+//
+//    @Autowired
+//    private ObjectMapper objectMapper;
+//
+//    @MockBean
+//    private LineService lineService;
+//
+//    @Test
+//    @DisplayName("최초의 노선(두개의 역)을 등록한다.")
+//    void create_line() throws Exception {
+//        final LineCreateRequest request = new LineCreateRequest("잠실역", "잠실새내역", "2호선", 3);
+//        final String content = objectMapper.writeValueAsString(request);
+//        final LineResponse3 lineResponse = new LineResponse3(1L, "2호선", "초록색");
+//        when(lineService.saveLine(any())).thenReturn(lineResponse);
+//
+//        mockMvc.perform(post("/lines")
+//                        .content(content)
+//                        .contentType(MediaType.APPLICATION_JSON_VALUE))
+//                .andDo(print())
+//                .andExpect(status().isCreated());
+//    }
+//}
